Features added : 	1. Categories for food in single restaurant view 	2. Show Rating, Duration Range, Price Range in single restaurant view UI Changes : 	1. Added icons to categories 	2. Changed single restaurant view 		i. Device Wide Picture at top 		ii. Restaurant Info and Category Selection 		iii. Menu

+++ lib/MapRoute.dart
@@ -31,7 +31,7 @@ class RouteInfo {
   double distance() {
     if (response != null) {
       var distance = response['routes'][0]['legs'][0]['distance'];
-      return distance;
+      return distance.toDouble();
     }
     return -1;
   }
